LayoutTests/imported/w3c/web-platform-tests/html/canvas/offscreen/text/2d.text.lang.default.worker-expected.txt


FAIL 2d assert_equals: ctx.lang === 'inherit' (got [undefined], expected inherit[string]) expected (string) "inherit" but got (undefined) undefined
